requirements.txt
discord.py==2.3.2
aiosqlite==0.20.0
pandas==2.2.2
python-dotenv==1.0.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
gspread==6.1.0
oauth2client==4.1.3
openpyxl==3.1.2